pyproject.toml
[project]
name = "restaurant-agent"
version = "0.1.0"
description = "State-driven restaurant reservation agent"
requires-python = ">=3.11"
authors = [
    { name = "AI Tutor System", email = "ai@example.com" }
]
dependencies = [
    "pydantic>=2.7",
    "jinja2>=3.1",
    "openai>=1.54",
    "python-dotenv>=1.0",
    "instructor>=1.4",
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["engine", "memory", "skills", "templates", "persistence", "shared"]
include-package-data = true

[tool.setuptools.package-data]
templates = ["**/*.j2"]
